requirements.txt
aiosqlite==0.19.0
alembic==1.10.3
fastapi==0.95.0
httpx==0.23.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.3.0
pytest_asyncio==0.21.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
sqlalchemy[asyncio]==2.0.9
uvicorn==0.21.1